Deal better with long tweets
Now we don't just truncate them wherever the 140-character limit is, but
only on spaces.
This is not only nice, but it's needed to avoid breaking links. We could
end up splitting a link, Twitter thinks it's a new valid link, it tries
to generate a short version of it, the short link is actually longer,
the tweet is still too long.